+++ src/app/pages/fna-bram/inputs/human-resources/recruitment-training/recruitment-training.component.ts
@@ -43,12 +43,18 @@ export class RecruitmentTrainingComponent implements OnInit {
   }
 
   delete(id) {
-    this.humanResourcesService.deleteHumanResourcesById(id).toPromise().then(async (res: any) => {
-      if (res) {
-        this.getAll();
-        this.alertService.activate('This record was deleted', 'Success Message');
+    this.alertService.activate('Are you sure want to delete?', 'Warning Message').then(result => {
+      if (result) {
+        this.humanResourcesService.deleteHumanResourcesById(id).toPromise().then(async (res: any) => {
+          if (res) {
+            this.getAll();
+            this.alertService.activate('This record was deleted', 'Success Message').then(result => {
+           
+            });
+          }
+        })
       }
-    })
+    });   
   }
 
   async getAll() {
